requirements.txt
aws-cdk-lib==2.231.0
aws-cdk.aws-bedrock-agentcore-alpha==2.232.1a0
constructs>=10.0.0,<11.0.0
